fix acquisition
