+++ src/qibocal/protocols/characterization/randomized_benchmarking/standard_rb.py
@@ -21,7 +21,7 @@ from .circuit_tools import (
 )
 from .data import RBData, RBType
 from .fitting import exp1B_func, fit_exp1B_func
-from .utils import number_to_str, random_clifford
+from .utils import number_to_str, random_clifford, resample_p0
 
 NPULSES_PER_CLIFFORD = 1.875
 
@@ -86,100 +86,19 @@ class StandardRBResult(Results):
     """Error bars for y."""
 
 
-def samples_to_p0(samples_list):
-    """Computes the probabilitiy of 0 from the list of samples.
-
-    Args:
-        samples_list (list or np.ndarray): 3d array with ``ncircuits`` rows containing
-            ``nshots`` lists with ``nqubits`` amount of ``0`` and ``1`` samples.
-            e.g. ``samples_list`` for 1 circuit, 3 shots and 2 qubits looks like
-            ``[[[0, 0], [0, 1], [1, 0]]]`` and ``p0=1/3``.
-
-    Returns:
-        list: list of probabilities corresponding to each row.
-    """
-
-    ground = np.array([0] * len(samples_list[0][0]))
-    return np.count_nonzero((samples_list == ground).all(axis=2), axis=1) / len(
-        samples_list[0]
-    )
+def layer_gen(nqubit_ids, seed):
+    """Returns a circuit with a random single-qubit clifford unitary."""
+    return random_clifford(nqubit_ids, seed)
 
 
-def resample_p0(data, sample_size=100, homogeneous: bool = True):
-    """Preforms parametric resampling of shots with binomial distribution
-        and returns a list of "corrected" probabilites.
-
-    Args:
-        data (list or np.ndarray): list of probabilities for the binomial distribution.
-        nshots (int): sample size for one probability distribution.
-
-    Returns:
-        list: resampled probabilities.
-    """
-    if homogeneous:
-        return np.apply_along_axis(
-            lambda p: samples_to_p0(
-                np.random.binomial(n=1, p=1 - p, size=(1, sample_size, len(p))).T
-            ),
-            0,
-            data,
-        )
-
-    resampled_data = []
-    for row in data:
-        resampled_data.append([])
-        for p in row:
-            samples_corrected = np.random.binomial(
-                n=1, p=1 - p, size=(1, sample_size, *p.shape)
-            ).T
-            resampled_data[-1].append(samples_to_p0(samples_corrected))
-    return resampled_data
-
-
-# def setup_scan(
-#     params: StandardRBParameters, qubits: Union[Qubits, list[QubitId]], nqubits: int
-# ) -> Iterable:
-#     """Returns an iterator of single-qubit random self-inverting Clifford circuits.
-
-#     Args:
-#         params (StandardRBParameters): Parameters of the RB protocol.
-#         qubits (dict[int, Union[str, int]] or list[Union[str, int]]):
-#             list of qubits the circuit is executed on.
-#         nqubits (int, optional): Number of qubits of the resulting circuits.
-#             If ``None``, sets ``len(qubits)``. Defaults to ``None``.
-
-#     Returns:
-#         Iterable: The iterator of circuits.
-#     """
-
-#     qubit_ids = list(qubits) if isinstance(qubits, dict) else qubits
-
-#     def make_circuit(depth):
-#         """Returns a random Clifford circuit with inverse of ``depth``."""
-
-#         # This function is needed so that the inside of the layer_circuit function layer_gen()
-#         # can be called for each layer of the circuit, and it returns a random layer of
-#         # Clifford gates. Could also be a generator, it just has to be callable.
-#         def layer_gen():
-#             """Returns a circuit with a random single-qubit clifford unitary."""
-#             return random_clifford(len(qubit_ids), params.seed)
-
-#         circuit = layer_circuit(layer_gen, depth)
-#         add_inverse_layer(circuit)
-#         add_measurement_layer(circuit)
-
-#         return embed_circuit(circuit, nqubits, qubit_ids)
-
-#     return map(make_circuit, params.depths * params.niter)
-
-
-def random_circuit(
+def random_circuits(
     depth: int,
-    params: StandardRBParameters,
-    qubits: Union[Qubits, list[QubitId]],
+    qubit_ids: Union[Qubits, list[QubitId]],
     nqubits: int,
+    niter,
+    seed,
 ) -> Iterable:
-    """Returns an iterator of single-qubit random self-inverting Clifford circuits.
+    """Returns single-qubit random self-inverting Clifford circuits.
 
     Args:
         params (StandardRBParameters): Parameters of the RB protocol.
@@ -192,25 +111,14 @@ def random_circuit(
         Iterable: The iterator of circuits.
     """
 
-    qubit_ids = list(qubits) if isinstance(qubits, dict) else qubits
-
-    def make_circuit(depth):
-        """Returns a random Clifford circuit with inverse of ``depth``."""
-
-        # This function is needed so that the inside of the layer_circuit function layer_gen()
-        # can be called for each layer of the circuit, and it returns a random layer of
-        # Clifford gates. Could also be a generator, it just has to be callable.
-        def layer_gen():
-            """Returns a circuit with a random single-qubit clifford unitary."""
-            return random_clifford(len(qubit_ids), params.seed)
-
-        circuit = layer_circuit(layer_gen, depth)
+    circuits = []
+    for _ in range(niter):
+        circuit = layer_circuit(layer_gen, depth, len(qubit_ids), seed)
         add_inverse_layer(circuit)
         add_measurement_layer(circuit)
-
-        return embed_circuit(circuit, nqubits, qubit_ids)
-
-    return map(make_circuit, depth * params.niter)
+        circuit = embed_circuit(circuit, nqubits, qubit_ids)
+        circuits.append(circuit)
+    return circuits
 
 
 def _acquisition(
@@ -241,7 +149,6 @@ def _acquisition(
     # For simulations, a noise model can be added.
     noise_model = None
     if params.noise_model:
-        # FIXME implement this check outside acquisition
         if platform and platform.name != "dummy":
             raise_error(
                 NotImplementedError,
@@ -264,71 +171,28 @@ def _acquisition(
     # 1. Set up the scan (here an iterator of circuits of random clifford gates with an inverse).
     nqubits = platform.nqubits if platform else max(qubits) + 1
     data = RBData(params=params.__dict__, depths=list(set(params.depths)))
+
     circuits = []
+    qubit_ids = list(qubits)
+    for depth in params.depths:
+        circuits_depth = random_circuits(
+            depth, qubit_ids, nqubits, params.niter, params.seed
+        )
+        circuits.extend(circuits_depth)
 
-    # # 2. Execute the scan.
+    # TODO: Check circuits being random properly
 
-    # samples = defaultdict(list)
-    # scan = setup_scan(params, qubits, nqubits)
-    # # Iterate through the scan and execute each circuit.
-    # for circuit in scan:
-    #     # The inverse and measurement gate don't count for the depth.
-    #     if noise_model is not None:
-    #         circuit = noise_model.apply(circuit)
-    #     circuits.append(circuit)
-
-    # multiple_sample =  backend.execute_circuits(circuits, nshots=params.nshots)
-    # for sample, circuit in zip(multiple_sample, circuits):
-    #     depth = (circuit.depth - 2) if circuit.depth > 1 else 0
-    #     sample = sample.samples()
-    #     # Every executed circuit gets a row where the data is stored.
-    #     # TODO: Try to get register qubit here
-    #     samples[depth].append(sample.tolist())
-
-    # new_samples = defaultdict(list)
-    # for qn, q in enumerate(qubits):
-    #     for d in params.depths:
-    #         for i in range(params.niter):
-    #             for n in range(params.nshots):
-    #                 new_samples[q, d].append(samples[d][i][n][qn])
-
-    # for key in list(new_samples.keys()):
-    #     data.register_qubit(
-    #         RBType,
-    #         key,
-    #         dict(
-    #             samples=new_samples[key[0], key[1]],
-    #         ),
-    #     )
-
-    for d in params.depths:
-        circuit = random_circuit(d, params, qubits, nqubits)
-        circuits.append(circuit)
     executed_circuits = backend.execute_circuits(circuits, nshots=params.nshots)
-
-    for executed_circuit in executed_circuits:
-        for qn, q in enumerate(qubits):
+    for executed_circuit, circuit in zip(executed_circuits, circuits):
+        depth = (circuit.depth - 2) if circuit.depth > 1 else 0
+        for nqubit, qubit_id in enumerate(qubits):
             data.register_qubit(
                 RBType,
-                (q, d),
+                (qubit_id, depth),
                 dict(
-                    samples=executed_circuit.samples()[qn],
+                    samples=executed_circuit.samples()[nqubit],
                 ),
             )
-
-    #     # signals = defaultdict(list)
-    #     # for key, value in samples.items():
-    #     #     if params.parallel:
-    #     #         signals[key] = samples_to_p0s(list(chain.from_iterable(value)))
-    #     #     else:
-    #     #         # TODO: Fix this
-    #     #         signals[key] = samples_to_p0(list(chain.from_iterable(value)))
-
-    # else:
-    #     # The signal here is the survival probability.
-    #     standardrb_data = RBData(
-    #         data.assign(signal=lambda x: samples_to_p0(x.samples.to_list()))
-    #     )
 
     # Store the parameters to display them later.
     data.params = params.__dict__
